Faster_whisper/requirements.txt
Flask
gunicorn
google-cloud-storage
faster-whisper